requirements.txt
streamlit
langchain
sentence-transformers
faiss-cpu
transformers
torch
PyPDF2
python-dotenv
langchain-community
pdfplumber
langchain_huggingface